mon: s/ENOSYS/ENOTSUP/
On Linux ENOTSUP is remapped
/usr/include/x86_64-linux-gnu/bits/errno.h

    /* Linux has no ENOTSUP error code.  */
    # define ENOTSUP EOPNOTSUPP

and should have different values on other operating systems. On Ubuntu
precise the string returned when translating the error value of ENOTSUP
or EOPNOTSUPP is always EOPNOTSUPP but on Ubuntu saucy it is always
ENOTSUP.

Replace ENOSYS with ENOTSUP because the expected semantic is very
similar and modify the test to not check on the string translation of
the error.

Rework the check_response shell function to optionaly check the return
code. The erasure coded pool size change test verifies the error message
only but not the error code.

Signed-off-by: Loic Dachary <loic@dachary.org>