mon: remove format argument from osd crush dump
The --format argument of the ceph cli is used to send the desired format
argument. The format argument is always part of the command sent to the
server. Adding it to the command description in MonCommand is not
necessary.

partially revert 6b93d586bca085885902a418097ca38f16502e8e
revert d03dfd212ded6b5e14a47add35cb39bae36b1595

Signed-off-by: Loic Dachary <loic@dachary.org>