mon: remove format argument from osd crush dump
The --format argument of the ceph cli is used to send the desired format
argument. The format argument is always part of the command sent to the
server. Adding it to the command description in MonCommand is not
necessary.

partially revert cec1893310005eca29186145de66391789211e70
revert fce4d6840448053b685ca3407b2a2680109d6df2

Signed-off-by: Loic Dachary <loic@dachary.org>