doc: update the operator create pool reference
Include erasure code information. A choice is made to display properties
that should be documented as part of the erasure code plugin
interface (erasure-code-k for instance). However, there only is one
plugin at the moment. Furthermore, erasure-code-k and erasure-code-m are
the two properties that the system administrator will want to tweak when
playing with erasure coded pools for the first time. It is probably
better to document them in a very visible place for this reason.

http://tracker.ceph.com/issues/7548 refs #7548

Signed-off-by: Loic Dachary <loic@dachary.org>