qa/workunits/rest/test.py: do not test 'osd thrash'
This wreaks havoc on our QA because it marks osds up and down and then
immediately after that we try to scrub and some osds are still down.  I
think it is not worth futzing with here since this is a debug/dev operation
and is covered by the CLI test.

Signed-off-by: Sage Weil <sage@inktank.com>