upgrade/dumpling-x: upgrade mon before osd
The MDS needs to see the TMAP2OMAP feature bit in the osdmap.  THe mons
have to encode the feature bits passed by the new osds in order for that
to happen.  Ergo, upgrade mons first.

This is only important for the MDS and for users who want to use erasure
coded pools or tiering without restarting the osds *again* after the mons
get upgraded.

Fixes: #7679
Signed-off-by: Sage Weil <sage@inktank.com>