mds: fix corner case of pushing inline data
Following sequence of events can happen.
 - Client releases an inode, queues cap release message.
 - A 'lookup' reply brings the same inode back, but the reply doesn't
   contain inline data because MDS didn't receive the cap release
   message and thought client already has up-to-data inline data.

To handle this corner case, the MDS need to add inline data to every
request reply or have a mechanism to allow client to request inline
data. The first approach is too expensive, the second one is tricky
to implement. This patch only makes sure client work well when inline
data is disabled.

Fixes: #7684
Signed-off-by: Yan, Zheng <zheng.z.yan@intel.com>