osd: replace pg_pool_t properties with erasure_code_profile
The generic map properties is only used for erasure code and should be
specialized. It is replaced by a string that can be looked up in
OSDMap::erasure_code_profiles to retrieve the key/value pairs that will
be interpreted by the erasure code plugin.

There is no upgrade path because properties is not found in any Ceph
release.

Signed-off-by: Loic Dachary <loic@dachary.org>