objecter: fixes / changes for homeless session
This is still not quite correct. Changed it so that we don't set the op
on the session if op->tid is 0. However, this means that we don't hold
any reference to this op. We probably need to have some way to generate
homeless-tids or something like that for the homeless ops. The orignal
code avoided that by using xlist.

Signed-off-by: Yehuda Sadeh <yehuda@inktank.com>